Se agregaron los requisitos extras, se realizaron correcciones y demas

+++ src/components/ItemDetailContainer/ItemDetailContainer.jsx
@@ -2,28 +2,35 @@ import { doc, getDoc } from "firebase/firestore";
 import { baseDeDatos } from "../../FireBase/config";
 import { useEffect, useState } from "react";
 import ItemDetail from "../ItemDetail/ItemDetail";
-import { useParams } from "react-router-dom";
-
+import { Link, useParams } from "react-router-dom";
 
 const ItemDetailContainer = () => {
-    const [item, setItem ] = useState(null);
+    const [item, setItem] = useState(null);
+    const [error, setError] = useState();
     const id = useParams().id;
 
     useEffect(() => {
-        const productoBaseDeDato = doc( baseDeDatos, "Productos", id ); 
+        const productoBaseDeDato = doc(baseDeDatos, "Productos", id);
 
         getDoc(productoBaseDeDato)
             .then((respuesta) => {
-                setItem({ ...respuesta.data(), id: respuesta.id });
+                if (respuesta.exists()) {
+                    setItem({ ...respuesta.data(), id: respuesta.id });
+                } else {
+                    setError("El producto ingresado no existe o no se encuenta disponible. Prueba buscando en nuestro catalogo");
+                }
             })
-    }, [id])
-    
+    }, [id]);
 
     return (
         <div>
-            {item && <ItemDetail item= {item} />}
+            {error ? (
+                <p>{error} <Link to="/productos">Catalogo</Link></p>
+            ) : (
+                item && <ItemDetail item={item} />
+            )}
         </div>
-    )
-}
+    );
+};
 
-export default ItemDetailContainer
+export default ItemDetailContainer;
